+++ dependencies.php
@@ -12,10 +12,23 @@ declare(strict_types=1);
 namespace Flextype\Plugin\Site;
 
 use Flextype\Plugin\Site\Controllers\SiteController;
+use Flextype\Plugin\Site\Models\Themes;
+
+/**
+ * Add themes service to Flextype container
+ */
+$flextype['themes'] = static function ($flextype) use ($app) {
+    return new Themes($flextype, $app);
+};
+
+/**
+ * Init themes
+ */
+$flextype['themes']->init($flextype, $app);
 
 /**
  * Add site controller to Flextype container
  */
-$flextype['SiteController'] = static function ($container) {
-    return new SiteController($container);
+$flextype['SiteController'] = static function ($flextype) use ($app) {
+    return new SiteController($flextype, $app);
 };